requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
redis
alembic
pydantic_settings
asyncpg
pytest
pytest-asyncio
httpx
passlib[bcrypt]==1.7.4
bcrypt<4.0
python-jose